pyproject.toml
[project]
name = "fastapiphotovideosharing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.135.3",
    "fastapi-users[sqlalchemy]>=15.0.5",
    "imagekitio>=5.2.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
    "uvicorn[standard]>=0.43.0",
]

[dependency-groups]
dev = [
    "autopep8>=2.3.2",
]
